Reestruturacao de codigo de index.js para Homescreen.js
