+++ public/index.html
@@ -15,7 +15,7 @@
 
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <section class="cardapio"></section>
+    <section class="root"></section>
     <footer>IES301 - Equipe 3 - 2021</footer>
   </body>
 </html>
