+++ src/index.css
@@ -37,7 +37,7 @@ code {
   height: 200px;
   background-color: #B9B7BD;
 }
-.cardapio{
+.root{
   margin-top: 10px !important;
   margin: 32px;
 }
